// packages/web/src/shared/api/generated/model/createScheduledPaymentDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */
import type { CreateScheduledPaymentDtoAmountMode } from "./createScheduledPaymentDtoAmountMode";
import type { CreateScheduledPaymentDtoScheduleKind } from "./createScheduledPaymentDtoScheduleKind";
import type { CreateScheduledPaymentDtoScheduleUnit } from "./createScheduledPaymentDtoScheduleUnit";

export interface CreateScheduledPaymentDto {
  /** @minLength 1 */
  name: string;
  amountMode: CreateScheduledPaymentDtoAmountMode;
  amount?: string;
  amountMin?: string;
  amountMax?: string;
  currency?: string;
  categoryId?: string;
  accountId?: string;
  assignedUserId?: string;
  scheduleKind: CreateScheduledPaymentDtoScheduleKind;
  /** @minimum 1 */
  scheduleInterval?: number;
  scheduleUnit?: CreateScheduledPaymentDtoScheduleUnit;
  /**
   * @minimum 1
   * @maximum 31
   */
  dueDay?: number;
  /**
   * @minimum 1
   * @maximum 12
   */
  dueMonth?: number;
  nextDueAt: string;
  timezone?: string;
  reminderDaysBefore?: number[];
  notifyTelegram?: boolean;
  notifyEmail?: boolean;
  notes?: string;
}