// packages/web/src/shared/api/generated/model/listScheduledPaymentsParams.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */

export type ListScheduledPaymentsParams = {
  take?: number;
  skip?: number;
  dueTo?: string;
  dueFrom?: string;
  categoryIds?: string[];
  accountIds?: string[];
  assignedUserIds?: string[];
  displayStatus?: string;
};