// packages/web/src/shared/api/generated/model/scheduledPaymentRecordResponseDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduledPaymentRecordResponseDtoStatus } from "./scheduledPaymentRecordResponseDtoStatus";

export interface ScheduledPaymentRecordResponseDto {
  id: string;
  scheduledPaymentId: string;
  workspaceId: string;
  /** @nullable */
  transactionId?: string | null;
  dueAt: string;
  /** @nullable */
  paidAt?: string | null;
  /** @nullable */
  skippedAt?: string | null;
  /** @nullable */
  amount?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  accountId?: string | null;
  /** @nullable */
  categoryId?: string | null;
  actionById: string;
  status: ScheduledPaymentRecordResponseDtoStatus;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}