// packages/web/src/shared/api/generated/model/updateScheduledPaymentDto.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateScheduledPaymentDtoAmountMode } from "./updateScheduledPaymentDtoAmountMode";
import type { UpdateScheduledPaymentDtoScheduleKind } from "./updateScheduledPaymentDtoScheduleKind";
import type { UpdateScheduledPaymentDtoScheduleUnit } from "./updateScheduledPaymentDtoScheduleUnit";

export interface UpdateScheduledPaymentDto {
  /** @minLength 1 */
  name?: string;
  amountMode?: UpdateScheduledPaymentDtoAmountMode;
  /** @nullable */
  amount?: string | null;
  /** @nullable */
  amountMin?: string | null;
  /** @nullable */
  amountMax?: string | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  categoryId?: string | null;
  /** @nullable */
  accountId?: string | null;
  /** @nullable */
  assignedUserId?: string | null;
  scheduleKind?: UpdateScheduledPaymentDtoScheduleKind;
  /** @minimum 1 */
  scheduleInterval?: number;
  /** @nullable */
  scheduleUnit?: UpdateScheduledPaymentDtoScheduleUnit;
  /**
   * @minimum 1
   * @maximum 31
   * @nullable
   */
  dueDay?: number | null;
  /**
   * @minimum 1
   * @maximum 12
   * @nullable
   */
  dueMonth?: number | null;
  nextDueAt?: string;
  timezone?: string;
  reminderDaysBefore?: number[];
  notifyTelegram?: boolean;
  notifyEmail?: boolean;
  /** @nullable */
  notes?: string | null;
}